refactor: cityDetail, city 패키지 통합
