+++ src/main/java/com/souf/soufwebsite/domain/city/repository/CityDetailRepository.java
@@ -1,6 +1,6 @@
-package com.souf.soufwebsite.domain.citydetail.repository;
+package com.souf.soufwebsite.domain.city.repository;
 
-import com.souf.soufwebsite.domain.citydetail.entity.CityDetail;
+import com.souf.soufwebsite.domain.city.entity.CityDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityDetailRepository extends JpaRepository<CityDetail, Long> {
